Added base template, improved answer parse.
We added the base.html template which contains all the common HTML
parts as well as placeholders for the parts differing in various page.
The common template has also the link to the CSS (we are using water
classless framework).

We have also improved the parse of answers because it was broken in the
sense that radio boxes were not working properly (due to options
having different names). All options (both radio and check boxes) have
now the same name and request.POST.getlist(key) is now used to retrieve
all selected options.