New views save and delete, minor changes.
Added save and delete views (with corresponding urls) used to save
a quiz during compilation and deleting it (currently not fully
implemented). The start view is also changed now it takes a parameter
which is the attempt id. the quiz layout is also changed now to be
more organized using Pico classless framework.
Attempt model class __str__ now shows also the id.