Added QuizApp app for the complete app.
The app named 'QuizApp' should be the final one in which we will
implement all the features, at least regarding the frontend. This app
will use the data model designed for the entire application and will
therefore have an extensive number of tables in the database.

During development, we will map this app in the quizapp/ url until
ready for alpha release and then move to the root (repositioning or
even removing the old quiz app which will be discontinued).