Changed question/show layout to be table based.
For a more compact view of the questions, all question/show routes
now show the list of questions in a table, one per row. The old
<article> based rendering is still available in the file
question_list_as_article and could be enabled by changing the included
template in the question_list block of the question_list.html template.