Refactored URLs and added JSON export.
We substituted the two urls `question/show/all` and `question/show/tags`
with a single one `question/show`, also adding the `subject` tag (which
should only be used once).

The submit button of the `question/show` view now downloads the JSON
file of the selected questions. This file can be used as input to the
python-quiz tool.