Integration of py-quiz tool.
With this commit, we started the integration of py-quiz CLI tool to the
django project. The source code of the tool is now within the python
path of django projects. Thus, accessing modules within the tools is
now straightforward within django apps.

This integration kept the CLI py-quiz.py module working as before
transferring to help the transition. In the future, the interface to
the tool should be a django app.