Implemented randomization on quiz generation.
When submitting from /question/show view, the random switch is now
implemented using the pyquiz module (former python-quiz). This module
is now runnable with `python -m` (not anymore with python py-quiz.py)
due to the requirements on relative imports.

Porting is not complete, but it still works as intended (a part from
different calling command) and also serves the QuizApp view.