Implemented Collection creation from question list
Along with the JSON download button, the /question/show view has now a
second button which allows the creation of a Collection model object
using the list of selected questions. Upon successful creation, the
landing page is the summary of the Collection with button to operate
on the Collection. At the moment, this button don't trigger any
operations.