[pyquiz] - Added support to question status.
We are adding support for a further attribute of the input JSON, which
is 'status', which allows defining how the program should behave on the
specific question. Currently the behavior is
- exclude question if "status": "hidden" unless --include-hidden is set
- include question otherwise.
Notice that, at the moment, the 'status' attribute is not passed to
Question object (hence it would be lost if it was to be re-rendered).