[pyquiz] - Added question id expansion.
This addition changes id's of questions during parsing. Notice that this
id changes are not saved on source file. More precisely, each id is
prepended with `n.`, where `n` is the index of source file.

This update partially fixes a bug in the correlation table printed when
verbosity is 2. The table uses sets to determine the intersection
of questions between tracks. Questions are defined with ids, but ids
don't need to be unique. By adding prefix based on file, uniqueness is
maintained even when equal id's are found in distinct files (note,
however, that the problem remains if identical id's appear on a same
file)