[quiz_app]: Added Django QuizGenWeb App.
The QuizGenWeb App is, essentially, a Django-based frontend for the
pyquiz command line App. This commit adds the app to the Django project.
We are currently implementing the frontend in a wizard style fashion.
Through wizard/... routes, we can create tests using files on the
filesystem as long as the path is present in the database (this is a
huge issue, since the database contains information about the local
filesystem structure). Future implementation of the App will have a
local storage for the files (which may be link to local files).