More on edit in gen/ application.
We added an intermediate page between file selection and question
editing pages. The workflow is now more 'wizard-like':
1. Select files
2. Show questions in selected files
3. Edit files

1. and 2. are functioning (although can be improved), while 3. is still
mocked.

---

We also added a button.html template that should be used every time a
button is needed. The template contains code to use several types of
buttons and can only be used with custom attributes (making it\
compatible with AlpineJS or other frontend frameworks).